
\name{grafGen_results}
\alias{grafGen_results}

\title{
    Reference data results
}
\description{
    The returned object from \code{\link{grafGen}} when analyzing
    the reference data.
}
\details{ An object of class "grafpop" containing the 
    \code{\link{grafGen}} results for each 
    of the 1011 genomes in the reference data. 
}
\examples{
    data(grafGen_results, package="GrafGen")
    grafGen_results
}
\keyword{data}
\seealso{ \code{\link{grafGen}} }
