
\name{Hpylori_prophage_data}
\alias{Hpylori_prophage_data}

\title{
    H. pylori prophage data
}
\description{ 
    SNP positions and allele frequencies for the reference data.
}
\details{
    A data frame containing the vertex and reference population 
    allele frequencies for the
    set of 22198 SNPs used in the analysis for H. pylori prophages.
    The SNPs were created using KHP30 as the reference genome.
    The total sample size was from a set of 80 strains.
}
\value{A data frame containing the reference SNPs.}
\examples{
    # Load data and view the first few rows
    data(Hpylori_prophage_data, package="GrafGen")
    Hpylori_prophage_data
}
\keyword{data}
