
\name{prophage_results}
\alias{prophage_results}

\title{
    Prophage data results
}
\description{
    The returned object from \code{\link{grafGen_prophage}} in the 
    analysis of the reference data.
}
\details{ An object of class "grafpop" containing the 
    \code{\link{grafGen_prophage}} results for each 
    of the 80 genomes in the reference data.
}
\value{ An object of class "grafpop".}
\examples{
    data(prophage_results, package="GrafGen")
    prophage_results
}
\keyword{data}
\seealso{ \code{\link{grafGen_prophage}} }
